REF fix

+++ app/Http/Controllers/TgBotController.php
@@ -46,7 +46,7 @@ class TgBotController extends Controller
         $action = mb_strtolower($action);
         $disable_web_page_preview = false;
         if (strripos($action, "start")) {
-            $link_id = str_replace('/start ', '', $action);
+            $link_id = (int)str_replace('/start ', '', $action);
             info('link_id: ' . print_r($link_id, 1));
             $action = '/start';
         } else {
